refactor: EditNameSectionをコンポーネント化
